+++ apps/backend/src/app/api/v1/auth/otp/sign-in/verification-code-handler.tsx
@@ -5,6 +5,7 @@ import { createVerificationCodeHandler } from "@/route-handlers/verification-cod
 import { VerificationCodeType } from "@prisma/client";
 import { UsersCrud } from "@stackframe/stack-shared/dist/interface/crud/users";
 import { signInResponseSchema, yupBoolean, yupNumber, yupObject, yupString } from "@stackframe/stack-shared/dist/schema-fields";
+import { createMfaRequiredError } from "../../mfa/sign-in/verification-code-handler";
 
 export const signInVerificationCodeHandler = createVerificationCodeHandler({
   metadata: {
@@ -23,6 +24,9 @@ export const signInVerificationCodeHandler = createVerificationCodeHandler({
   data: yupObject({
     user_id: yupString().required(),
     is_new_user: yupBoolean().required(),
+  }),
+  method: yupObject({
+    email: yupString().email().required(),
   }),
   response: yupObject({
     statusCode: yupNumber().oneOf([200]).required(),
@@ -43,6 +47,22 @@ export const signInVerificationCodeHandler = createVerificationCodeHandler({
     });
   },
   async handler(project, { email }, data) {
+    const projectUserBefore = await prismaClient.projectUser.findUniqueOrThrow({
+      where: {
+        projectId_projectUserId: {
+          projectId: project.id,
+          projectUserId: data.user_id,
+        },
+      },
+    });
+    if (projectUserBefore.requiresTotpMfa) {
+      throw await createMfaRequiredError({
+        project,
+        isNewUser: data.is_new_user,
+        userId: projectUserBefore.projectUserId,
+      });
+    }
+
     const projectUser = await prismaClient.projectUser.update({
       where: {
         projectId_projectUserId: {