+++ packages/stack/src/components-page/account-settings.tsx
@@ -1,13 +1,20 @@
 'use client';
 
-import React from 'react';
-import { useUser } from '..';
+import React, { useEffect } from 'react';
+import { CurrentUser, Project, useStackApp, useUser } from '..';
 import { PredefinedMessageCard } from '../components/message-cards/predefined-message-card';
 import { UserAvatar } from '../components/elements/user-avatar';
 import { useState } from 'react';
 import { FormWarningText } from '../components/elements/form-warning';
 import { getPasswordError } from '@stackframe/stack-shared/dist/helpers/password';
 import { Button, Card, CardContent, CardFooter, CardHeader, Container, Input, Label, PasswordInput, Typography, cn } from '@stackframe/stack-ui';
+import { generateRandomValues } from '@stackframe/stack-shared/dist/utils/crypto';
+import { TOTPController, createTOTPKeyURI } from "oslo/otp";
+import * as QRCode from 'qrcode';
+import { throwErr } from '@stackframe/stack-shared/dist/utils/errors';
+import { useAsyncCallback } from '@stackframe/stack-shared/dist/hooks/use-async-callback';
+import { runAsynchronously, runAsynchronouslyWithAlert } from '@stackframe/stack-shared/dist/utils/promises';
+import { set } from 'react-hook-form';
 
 function SettingSection(props: {
   title: string,
@@ -107,14 +114,14 @@ function EmailVerificationSection() {
       }}
     >
       {user?.primaryEmailVerified ?
-        <Typography variant='success'>Your email has been verified</Typography> :
-        <Typography variant='destructive'>Your email has not been verified</Typography>}
+        <Typography variant='success'>Your email has been verified.</Typography> :
+        <Typography variant='destructive'>Your email has not been verified.</Typography>}
     </SettingSection>
   );
 }
 
 function PasswordSection() {
-  const user = useUser();
+  const user = useUser({ or: "throw" });
   const [oldPassword, setOldPassword] = useState<string>('');
   const [oldPasswordError, setOldPasswordError] = useState<string>('');
   const [newPassword, setNewPassword] = useState<string>('');
@@ -123,7 +130,7 @@ function PasswordSection() {
   const [repeatNewPasswordError, setRepeatNewPasswordError] = useState<string>('');
   const [passwordChanged, setPasswordChanged] = useState(false);
 
-  if (!user?.hasPassword) {
+  if (!user.hasPassword) {
     return null;
   }
 
@@ -211,15 +218,105 @@ function PasswordSection() {
   );
 }
 
+function MfaSection() {
+  const project = useStackApp().useProject();
+  if (project.config.oauthProviders.length !== 0 || project.config.magicLinkEnabled) {
+    // TODO next-release support MFA for OAuth and magic link
+    return null;
+  }
+
+  const user = useUser({ or: "throw" });
+  const [generatedSecret, setGeneratedSecret] = useState<Uint8Array | null>(null);
+  const [qrCodeUrl, setQrCodeUrl] = useState<string | null>(null);
+  const [mfaCode, setMfaCode] = useState<string>("");
+  const [isMaybeWrong, setIsMaybeWrong] = useState(false);
+  const isEnabled = user.isMultiFactorRequired;
+
+  const [handleSubmit, isLoading] = useAsyncCallback(async () => {
+    await user.update({
+      totpMultiFactorSecret: generatedSecret,
+    });
+    setGeneratedSecret(null);
+    setQrCodeUrl(null);
+    setMfaCode("");
+  }, [generatedSecret, user]);
+
+  useEffect(() => {
+    setIsMaybeWrong(false);
+    runAsynchronouslyWithAlert(async () => {
+      if (generatedSecret && await new TOTPController().verify(mfaCode, generatedSecret)) {
+        await handleSubmit();
+      }
+      setIsMaybeWrong(true);
+    });
+  }, [mfaCode, generatedSecret, handleSubmit]);
+
+  return (
+    <SettingSection
+      title='Multi-factor Authentication'
+      desc='Secure your account with an additional layer of security.'
+      buttonVariant='secondary'
+      buttonText={isEnabled ? 'Disable' : (generatedSecret ? 'Cancel' : 'Enable')}
+      onButtonClick={async () => {
+        if (isEnabled) {
+          await user.update({
+            totpMultiFactorSecret: null,
+          });
+        } else if (!generatedSecret) {
+          const secret = generateRandomValues(new Uint8Array(20));
+          setQrCodeUrl(await generateTotpQrCode(project, user, secret));
+          setGeneratedSecret(secret);
+        } else {
+          setGeneratedSecret(null);
+          setQrCodeUrl(null);
+          setMfaCode("");
+        }
+      }}
+    >
+      {isEnabled ? (
+        <Typography variant="success">Multi-factor authentication is currently enabled.</Typography>
+      ) : (
+        generatedSecret ? (
+          <div className='flex flex-col gap-4 items-center'>
+            <Typography>Scan this QR code with your authenticator app:</Typography>
+            <img width={200} height={200} src={qrCodeUrl ?? throwErr("TOTP QR code failed to generate")} alt="TOTP multi-factor authentication QR code" />
+            <Typography>Then, enter your six-digit MFA code:</Typography>
+            <Input
+              value={mfaCode}
+              onChange={(e) => {
+                setIsMaybeWrong(false);
+                setMfaCode(e.target.value);
+              }}
+              placeholder="123456"
+              maxLength={6}
+              disabled={isLoading}
+            />
+            {isMaybeWrong && mfaCode.length === 6 && (
+              <Typography variant="destructive">Incorrect code. Please try again.</Typography>
+            )}
+          </div>
+        ) : (
+          <Typography variant="destructive">Multi-factor authentication is currently disabled.</Typography>
+        )
+      )}
+    </SettingSection>
+  );
+}
+
+async function generateTotpQrCode(project: Project, user: CurrentUser, secret: Uint8Array) {
+  const uri = createTOTPKeyURI(project.displayName, user.primaryEmail ?? user.id, secret);
+  return await QRCode.toDataURL(uri) as any;
+}
+
 function SignOutSection() {
-  const user = useUser();
+  const user = useUser({ or: "throw" });
   return (
     <SettingSection
       title='Sign out'
       desc='Sign out of your account on this device.'
       buttonVariant='secondary'
       buttonText='Sign Out'
-      onButtonClick={() => user?.signOut()}
+      onButtonClick={() => user.signOut()}
     >
     </SettingSection>
   );
@@ -241,6 +338,7 @@ export function AccountSettings({ fullPage=false }: { fullPage?: boolean }) {
       <ProfileSection />
       <EmailVerificationSection />
       <PasswordSection />
+      <MfaSection />
       <SignOutSection />
     </div>
   );