+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @stackframe/stack-backend
 
+## 2.5.18
+
+### Patch Changes
+
+- Multi-factor authentication
+- Updated dependencies
+  - @stackframe/stack-emails@2.5.18
+  - @stackframe/stack-shared@2.5.18
+
 ## 2.5.17
 
 ### Patch Changes