+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,13 @@
 # e2e-tests
 
+## 2.5.18
+
+### Patch Changes
+
+- Multi-factor authentication
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.18
+
 ## 2.5.17
 
 ### Patch Changes